+++ src/AnomalyDetectors/LODA.php
@@ -4,7 +4,6 @@ namespace Rubix\ML\AnomalyDetectors;
 
 use Rubix\ML\Online;
 use Rubix\ML\Learner;
-use Rubix\ML\Ranking;
 use Rubix\Tensor\Matrix;
 use Rubix\Tensor\Vector;
 use Rubix\ML\Persistable;